refactor(#419): hoist typed-acceptance evaluation to the cycle-handler base
Moves _evaluate_typed_acceptance + evaluator-error escalation from
DevelopmentDevelopHandler to _CycleTaskHandler so every cycle handler whose
plan task carries typed criteria enforces them identically — no copy-paste
per role. The hoisted seam coerces wire-shape criteria via
split_acceptance_criteria (#420), so M1.3 now actually executes on
dispatched tasks; regression tests pin the dict-criteria path.
