+++ src/squadops/capabilities/handlers/cycle/develop.py
@@ -5,9 +5,7 @@ Split from cycle_tasks.py (#152).
 from __future__ import annotations
 
 import logging
-import tempfile
 import time
-from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
 from squadops.capabilities.dev_capabilities import get_capability
@@ -16,9 +14,6 @@ from squadops.capabilities.handlers.base import (
     HandlerResult,
 )
 from squadops.capabilities.handlers.prompt_guard import _guard_prompt_size
-from squadops.cycles.acceptance_checks import CheckOutcome, get_check
-from squadops.cycles.implementation_plan import TypedCheck
-from squadops.cycles.patch_verification import materialize_artifacts
 from squadops.llm.exceptions import LLMError
 from squadops.llm.models import ChatMessage
 
@@ -177,158 +172,9 @@ class DevelopmentDevelopHandler(_CycleTaskHandler):
             summary="; ".join(summary_parts) or "All checks passed",
         )
 
-    # ---- SIP-0092 M1.3 typed acceptance helpers --------------------------
-
-    async def _evaluate_typed_acceptance(
-        self,
-        inputs: dict[str, Any],
-        artifacts: list[dict],
-        checks: list[dict],
-        missing: list[str],
-        typed_error_counts: dict[str, int],
-    ) -> None:
-        """Evaluate typed acceptance criteria, mutating ``checks`` and ``missing`` in place.
-
-        Implements RC-9 (severity × status blocking matrix), RC-9a (error-vs-
-        failed wording distinction), RC-9b (per-criterion error count with
-        2-strikes escalation), RC-12a (skipped-not-error for unset stack).
-        """
-        criteria = inputs.get("acceptance_criteria", [])
-        typed_criteria = [c for c in criteria if isinstance(c, TypedCheck)]
-        prose_criteria = [c for c in criteria if not isinstance(c, TypedCheck)]
-
-        # Prose strings stay informational, evidence-only — same as Rev 1's
-        # included_in_evidence behavior. They never block.
-        if prose_criteria:
-            checks.append(
-                {
-                    "check": "acceptance_criteria_prose",
-                    "criteria": prose_criteria,
-                    "evaluation": "included_in_evidence",
-                    "passed": True,
-                }
-            )
-
-        if not typed_criteria:
-            return
-
-        resolved_config = inputs.get("resolved_config", {})
-        typed_acceptance_enabled = resolved_config.get("typed_acceptance", True)
-        command_acceptance_enabled = resolved_config.get("command_acceptance_checks", True)
-        stack = resolved_config.get("stack")
-
-        with tempfile.TemporaryDirectory(prefix="squadops-typed-acc-") as tmpdir_str:
-            workspace_root = Path(tmpdir_str)
-            self._materialize_artifacts(artifacts, workspace_root)
-
-            for check_index, criterion in enumerate(typed_criteria):
-                outcome = await self._evaluate_typed_check(
-                    criterion,
-                    workspace_root,
-                    stack=stack,
-                    typed_acceptance_enabled=typed_acceptance_enabled,
-                    command_acceptance_enabled=command_acceptance_enabled,
-                )
-                check_record = {
-                    "check": f"acceptance:{criterion.check}",
-                    "severity": criterion.severity,
-                    "params": criterion.params,
-                    "description": criterion.description,
-                    "status": outcome.status,
-                    "actual": outcome.actual,
-                    "reason": outcome.reason,
-                    # `passed` flag for compatibility with the legacy
-                    # all-checks-pass aggregator: only severity=error AND
-                    # blocking status counts as not-passed.
-                    "passed": not (
-                        criterion.severity == "error" and outcome.status in {"failed", "error"}
-                    ),
-                    # Issue #114: identity fields for downstream trigger
-                    # composition and per-cycle evaluation persistence.
-                    # task_index is None for tasks not driven by an
-                    # implementation_plan (legacy monolithic flow).
-                    "task_index": inputs.get("subtask_index"),
-                    "check_index": check_index,
-                }
-                checks.append(check_record)
-
-                # Issue #83: per-check observability. Without these the M1.3
-                # path is invisible to operators — see issue body for context.
-                blocking = criterion.severity == "error" and outcome.status in {"failed", "error"}
-                log_fn = logger.info if blocking else logger.debug
-                log_fn(
-                    "typed_acceptance_check subtask=%s check=%s severity=%s status=%s blocking=%s reason=%s",
-                    inputs.get("subtask_index"),
-                    criterion.check,
-                    criterion.severity,
-                    outcome.status,
-                    blocking,
-                    outcome.reason or "",
-                )
-
-                # RC-9: severity AND status are independent. Only error+blocking missions.
-                if criterion.severity != "error":
-                    continue
-                if outcome.status == "failed":
-                    # RC-9a: app-incompleteness wording.
-                    label = criterion.description or criterion.check
-                    missing.append(f"acceptance:{label}")
-                elif outcome.status == "error":
-                    fp = criterion.fingerprint()
-                    prior = typed_error_counts.get(fp, 0)
-                    if prior < 2:
-                        # RC-9a: evaluator-error wording, distinct from app-incomplete.
-                        missing.append(f"evaluator-error:{criterion.check}: {outcome.reason}")
-                    else:
-                        self._escalate_persistent_evaluator_error(criterion, outcome)
-                    typed_error_counts[fp] = prior + 1
-                # status in {passed, skipped} never blocks.
-
-    # Shared with the executor-side patch verification (#389) — one
-    # materializer, one safety policy.
-    _materialize_artifacts = staticmethod(materialize_artifacts)
-
-    @staticmethod
-    async def _evaluate_typed_check(
-        criterion: TypedCheck,
-        workspace_root: Path,
-        *,
-        stack: str | None,
-        typed_acceptance_enabled: bool,
-        command_acceptance_enabled: bool,
-    ) -> CheckOutcome:
-        """Dispatch a typed criterion to its registered evaluator, honoring config gates."""
-        if not typed_acceptance_enabled:
-            return CheckOutcome.skipped(reason="typed_acceptance_disabled")
-        if criterion.check == "command_exit_zero" and not command_acceptance_enabled:
-            return CheckOutcome.skipped(reason="command_acceptance_checks_disabled")
-        try:
-            evaluator = get_check(criterion.check)
-        except KeyError:
-            # Should not happen — parser already enforces vocabulary — but
-            # treat as evaluator-error rather than crashing the cycle.
-            return CheckOutcome.error(reason="no_evaluator_registered")
-        return await evaluator.evaluate(criterion.params, workspace_root, stack=stack)
-
-    @staticmethod
-    def _escalate_persistent_evaluator_error(criterion: TypedCheck, outcome: CheckOutcome) -> None:
-        """RC-9b: surface a persistent evaluator error outside the self-eval feedback loop.
-
-        Logged at WARNING with structured fields; the correction protocol
-        and operator-facing surfaces consume the log. A first-class
-        escalation channel is a follow-up if/when the prompt-feedback
-        suppression proves insufficient.
-        """
-        logger.warning(
-            "typed_check_evaluator_error_escalated",
-            extra={
-                "check": criterion.check,
-                "severity": criterion.severity,
-                "fingerprint": criterion.fingerprint(),
-                "reason": outcome.reason,
-                "actual": outcome.actual,
-            },
-        )
+    # SIP-0092 M1.3 typed-acceptance evaluation is inherited from
+    # _CycleTaskHandler (#419/#420) — one seam for every cycle handler
+    # whose plan task carries typed acceptance criteria.
 
     def _validate_monolithic(
         self,
